requirements.txt
Flask
numpy
Pillow
werkzeug
trimesh